abstract-user/pyproject.toml
[tool.poetry]
name = "testproject"
version = "0.1.0"
description = ""
authors = ["Lauri Hintsala <lauri.hintsala@verkkopaja.fi>"]

[tool.poetry.dependencies]
python = "^3.9"
strawberry-graphql-django = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
